fix: import TransactionsOnChain section Framer Motion's components on client-side
