+++ src/components/DataRoom/TransactionsOnChain/index.tsx
@@ -1,13 +1,11 @@
 import { Typography } from '@mui/material'
+import dynamic from 'next/dynamic'
 import type { BaseBlock } from '@/components/Home/types'
-import Counter from './Counter'
-import { useScroll, useTransform, motion } from 'framer-motion'
-import React, { useRef } from 'react'
-import css from './styles.module.css'
-import { useCounterScroll } from './useCounterScroll'
 import LinksWrapper from '../LinksWrapper'
-import { useIsMediumScreen } from '@/hooks/useMaxWidth'
 import { useSafeDataRoomStats } from '@/hooks/useSafeDataRoomStats'
+import css from './styles.module.css'
+
+const CounterContainer = dynamic(() => import('./CounterContainer'))
 
 const TRANSACTIONS_ON_CHAIN_PERCENTAGE_FALLBACK = 1.75
 
@@ -30,43 +28,6 @@ const TransactionsOnChain = ({ text, link }: BaseBlock) => {
         </div>
       </div>
     </div>
-  )
-}
-
-const CounterContainer = ({ percentage }: { percentage: number }) => {
-  const targetScrollRef = useRef<HTMLDivElement>(null)
-
-  const { scrollYProgress } = useScroll({
-    target: targetScrollRef,
-    offset: ['start end', 'end start'],
-  })
-
-  const isMobile = useIsMediumScreen()
-  const opacity = useTransform(scrollYProgress, [0, 0.5], [0.1, 1])
-  const yTransform = useTransform(scrollYProgress, [0, 0.5], isMobile ? [200, 0] : [600, 0])
-  const value = useCounterScroll(scrollYProgress, percentage)
-
-  return (
-    <motion.div
-      style={{
-        opacity,
-        y: yTransform,
-      }}
-      className={css.counterContainer}
-      ref={targetScrollRef}
-    >
-      <div className={css.box1}>
-        <Counter value={value} />
-      </div>
-
-      <div className={css.box2}>
-        <Counter value={value} />
-      </div>
-
-      <div>
-        <Counter value={value} />
-      </div>
-    </motion.div>
   )
 }
