+++ src/content/dataroom.json
@@ -44,7 +44,6 @@
   {
     "component": "DataRoom/VolumeTransferred",
     "title": "with all time volume<br/>transferred",
-    "text": "and counting...",
     "link": {
       "title": "Dune dashboard",
       "href": "https://dune.com/safe/all"
